Q16

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,7 +9,7 @@ int main(void)
 {
 int i;
 /* long int pf, af, f, f1, f2, f3;*/
-long double f, pf, af ;
+unsigned long f, pf, af ;
 
 pf = 1;
 af = 2;
